pyproject.toml
[project]
name = "glitchygames"
version = "1.0.0"
description = "Glitchy Games and Tools for Pygame and Python 3"
authors = [
  { name = "Terry Simons", email = "terry.simons@gmail.com" },
  { name = "Rich Saupe", email = "sabadam32@gmail.com" },
]
maintainers = [
  { name = "Terry Simons", email = "terry.simons@gmail.com" },
  { name = "Rich Saupe", email = "sabadam32@gmail.com" },
]
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.13,<4.0"
classifiers = [
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
]
keywords = ["pygame", "game-engine", "pixel-art", "sprite-editor"]
dependencies = [
  "tomli-w",
  "pygame-ce",
  "cookiecutter",
  "pyperclip",
  "aisuite[all]>=0.1.14",
  "numpy",
  "miniaudio",
  "speechrecognition",
  "bokeh",
  "backoff",
  "fastapi",
  "uvicorn",
  "pydantic",
  "httpx",
  "apng",
  "pillow",
  "deptry>=0.25.1",
  "vulture>=2.15",
  "pip-audit>=2.10.0",
  "hypothesis>=6.151.9",
]

[project.urls]
"Bug Tracker" = "https://github.com/terrysimons/glitchygames/issues/"
"Homepage" = "https://glitchygames.ai/"
"Documentation" = "https://glitchygames.ai/documentation/"
"Repository" = "https://github.com/terrysimons/glitchygames"

[project.scripts]
bitmappy = "glitchygames.tools.bitmappy:main"
glitchygames-server = "glitchygames.api.main:run"
glitchygames-client = "glitchygames.api.client:run"

[project.optional-dependencies]
docs = [
  "mkdocs",
  "mkdocstrings[crystal,python]",
  "mkdocs-material",
  "mkdocs-gen-files",
  "mkdocs-literate-nav",
  "mkdocs-section-index",
  "pymdown-extensions",
  "mkdocs-autorefs",
  "mkdocs-macros-plugin",
  "mkdocstrings-python",
]
dev = [
  "nox[uv]",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "pytest-benchmark",
  "pytest-mock",
  "bandit",
  "safety",
  "ruff",
  "pre-commit",
  "basedpyright",
  "ty",
  "pillow",
]
api = []

[tool.hatch.build.targets.wheel]
packages = ["glitchygames"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Exclude examples directory from test collection
norecursedirs = [
  "examples",
  "build",
  "dist",
  "*.egg-info",
  ".git",
  ".hypothesis",
  "__pycache__",
]
addopts = "-n auto --cov=glitchygames --cov-fail-under=80 --log-cli-level=DEBUG --junitxml=test-results/junit.xml"
filterwarnings = [
  "error",
  # Third-party: speechrecognition depends on aifc, removed from stdlib in 3.13
  "ignore:aifc was removed:DeprecationWarning",
  # Third-party: docstring_parser uses deprecated ast nodes (via aisuite)
  "ignore:ast\\..* is deprecated:DeprecationWarning",
  # pytest-benchmark warns when xdist is active (benchmarks auto-disabled)
  "ignore:Benchmarks are automatically disabled:pytest_benchmark.logger.PytestBenchmarkWarning",
]
# Note: All tests run with no_unhandled_events=True to catch unhandled events as bugs

[tool.coverage.run]
source = ["glitchygames"]
branch = true
parallel = true
omit = [
  "*/tests/*",
  "*/examples/*",
  "*/build/*",
  "*/dist/*",
  "*/__pycache__/*",
  "*/migrations/*",
]

[tool.coverage.report]
show_missing = true

exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "if settings.DEBUG",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == .__main__.:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
  "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
preview = true
line-length = 100
output-format = "grouped"
src = ["glitchygames", "tests", "scripts"]
include = [
  "glitchygames/*.py",
  "tests/**/*.py",
  "scripts/**/*.py",
  "noxfile.py",
]
target-version = "py313"

[tool.ruff.lint]
task-tags = ["HACK", "TODO", "FIXME", "XXX", "NOTE", "BUG", "OPTIMIZE", "WIP"]
select = ["ALL"]
ignore = [
  # --- Previously ignored ---
  #"ASYN",
  "FIX002",  # Line contains TODO, consider resolving the issue
  "SLF001",  # Private member accessed (Needed for some annotations, like argparse._ArgumentGroup)
  "S311",    # Standard pseudo-random generators are not suitable for security/cryptographic purposes.
  "COM812",  # Trailing comma missing
  "G004",    # Logging statement uses f-string
  "ISC002",  # Use of implicit string concatenation
  "ERA001",  # Found commented-out code
  "PLR0911", # Too many return statements
  "C901",    # Function is too complex
  "EM102",   # Exception must not use an f-string literal, assign to variable first
  "TRY003",  # Avoid specifying long messages outside the exception class
  "ARG002",  # Unused method argument
  "TD003",   # Missing issue link on the line following this TODO
  "PLR0913", # Too many arguments to function call
  "TD002",   # Missing author in TODO
  "PLR0917", # Too many positional arguments
  "PLR0904", # Too many public methods
  "CPY001",  # Missing copyright notice at top of file
  "B019",    # Use of `functools.lru_cache` or `functools.cache` on methods can lead to memory leaks
  "EM101",   # Exception must not use a string literal, assign to variable first
  "ISC001",  # Use of implicit string concatenation (this conflicts with ruff format)
  "D203",    # Incorrect blank line before class (conflicts with D211)
  "D213",    # Multi-line summary second line (conflicts with D212)
  "TRY300",  # Consider moving this statement to an `else` block
  "PLR6301", # Method could be a function/staticmethod
  "PLC0415", # import-outside-top-level (intentional: circular imports, optional deps, platform-specific)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "PLR2004",
  "ANN",
  "ARG001",
  "ARG005",
  "PLC2701",
  "PLC2801", # Tests may call dunder methods directly to verify they work
  "D102",    # Test method names are self-documenting
  "DOC201",  # Test helper return docs unnecessary
  "RET504",  # Explicit assignment before return aids test readability
  "F841",
  "B007",
  "S101",
]

[tool.ruff.format]
docstring-code-format = true
quote-style = "single"
line-ending = "lf"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = false

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["glitchygames"]

## basedpyright config lives in pyrightconfig.json (supports per-environment overrides)

## ty (Astral's type checker) configuration
[tool.ty.environment]
python-version = "3.13"
extra-paths = ["scripts"]

[tool.ty.src]
include = ["glitchygames", "tests", "scripts"]

[tool.ty.rules]
# type: ignore comments are shared with basedpyright — don't flag them as unused
unused-type-ignore-comment = "ignore"

[tool.ty.terminal]
output-format = "full"
error-on-warning = false


[tool.deptry.per_rule_ignores]
# numpy is used indirectly via pygame.surfarray (array3d/array4d) which requires it at runtime
# Dev/build/lint/docs tools are invoked via CLI, not imported in Python code
DEP002 = [
  "numpy",
  "deptry",
  "mkdocs",
  "mkdocstrings",
  "mkdocs-material",
  "mkdocs-literate-nav",
  "mkdocs-section-index",
  "pymdown-extensions",
  "mkdocs-autorefs",
  "mkdocs-macros-plugin",
  "mkdocstrings-python",
  "pytest-cov",
  "pytest-xdist",
  "pytest-benchmark",
  "pytest-mock",
  "bandit",
  "safety",
  "ruff",
  "pre-commit",
  "basedpyright",
  "ty",
  "vulture",
  "pip-audit",
  "hypothesis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
